refactor: 상품 목록 리스트 섹션 컴포넌트 분리
